+++ src/en/silentquill/res/mipmap-hdpi/ic_launcher.png
