+++ src/en/silentquill/res/mipmap-mdpi/ic_launcher.png
