+++ src/en/silentquill/res/mipmap-xxxhdpi/ic_launcher.png
